Consolidate AccessPanelLinks behavior into MarcLinks.
